fix move tool in arbitrary orientation
