+++ frontend/javascripts/oxalis/controller/scene_controller.ts
@@ -13,6 +13,7 @@ import type {
   Vector3,
 } from "oxalis/constants";
 import constants, {
+  OrthoBaseRotations,
   OrthoViews,
   OrthoViewValuesWithoutTDView,
   TDViewDisplayModeEnum,
@@ -72,13 +73,6 @@ THREE.Mesh.prototype.raycast = acceleratedRaycast;
 
 const CUBE_COLOR = 0x999999;
 const LAYER_CUBE_COLOR = 0xffff99;
-
-export const OrthoBaseRotations = {
-  [OrthoViews.PLANE_XY]: new THREE.Euler(Math.PI, 0, 0),
-  [OrthoViews.PLANE_YZ]: new THREE.Euler(Math.PI, (1 / 2) * Math.PI, 0),
-  [OrthoViews.PLANE_XZ]: new THREE.Euler((-1 / 2) * Math.PI, 0, 0),
-  [OrthoViews.TDView]: new THREE.Euler(Math.PI / 4, Math.PI / 4, Math.PI / 4),
-};
 
 const getVisibleSegmentationLayerNames = reuseInstanceOnEquality((storeState: WebknossosState) =>
   getVisibleSegmentationLayers(storeState).map((l) => l.name),
